feat: restructure routes to separate folders and isolate experiments
