+++ packages/simulator/web/routes/measureStations.js
@@ -1,0 +1,13 @@
+const { toArray } = require('rxjs')
+
+const register = (experiment, socket) => {
+  return [
+    experiment.measureStations.pipe(toArray()).subscribe((measureStations) => {
+      socket.emit('measureStations', measureStations)
+    }),
+  ]
+}
+
+module.exports = {
+  register,
+}
